Improve check for Python package installation
Not all packages can be imported by their PyPI package name (e.g.
`python-dateutil` is imported as `dateutil`) so use the Py3.8+
importlib `metadata` component to check for the package by its
installed name. Also ensure that the package is not imported by the
check, which can cause side effects.